moved from clang-by-default to gcc
I was having some problems with projects that expected GCC to be the
compiler used to build them, but didn't explicitly specify it. The move
to Clang was made years ago when I read somewhere it started achieving
better results than GCC. I'm told that's no longer the case, but alas
GCC is still nearly omnipresent at this point.