changed workspace name and updated vim modeline
`main` was a weird name anyway. With the new vim modeline and my neovim
configuration, editing `xmonad.hs` will set `makeprg` so that whenever I
write a file under `.xmonad`, XMonad will try to recompile itself.

Upon writing the above paragraph I noticed this is probably not what I
want when doing a lot of rewriting, but hey, if it comes to that I'll
just disable it and everything should be fine again.