xmonad: send discord to web workspace
Yes, I installed Discord on this computer. No, I'm not proud of this,
okay? I was extremely hesitant to even make this commit as I thought
this wouldn't stay on my computer for too long, but the fact that the
repository had unstaged changes got increasingly annoying over time.

I won't push this to GitHub or GitLab anytime soon. If I do, well, rip.
I hope to remove this commit some time in the future before I do that.

Screw Discord, I still hate it! ;-;